// libs/help-desk/src/lib/models/employee-info.ts
export interface EmployeeInfo {
  status: string;
  cif: string;
  fullName: string;
  dateOfBirth: string;
  gender: string;
  maritalStatus: string;
  phoneNumber: string;
  team: string;
  email: string;
  skype: string;
  office: string;
  nationality: string;
  languages: string[] | string;
  jobTitle: string;
  joinDate: string;
  major: string[] | string;
  institute: string[] | string;
  facebook: string;
  instagram: string;
}